requirements.txt
Django == 4.0.6
djangorestframework==3.13.1
PyMySQL==1.0.2
django-cors-headers==3.13.0
django-storages==1.13.1